+++ src/index.html
@@ -143,6 +143,5 @@
       </section>
     </main>
 
-    <script src="./js/bundle.js"></script>
   </body>
 </html>
