+++ src/js/data/cards.json
@@ -1,0 +1,81 @@
+{
+  "cards": [
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": ",#repeat",
+      "cardColor": "black"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "yellow"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "blue"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "green"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "pink"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "pink"
+    },
+    {
+      "cardText": "It is example of repeating task. It marks by wave.",
+      "cardImage": "img/sample-img.jpg",
+      "cardDeadline": {
+        "date": "23 September",
+        "time": "11:15 PM"
+      },
+      "cardRepeat": "No",
+      "cardHashtag": "#repeat",
+      "cardColor": "pink"
+    }
+  ]
+}
