+++ src/js/modules/_board.js
@@ -1,147 +1,9 @@
-// DATA
-const dataList = {
-  filters: [
-    {
-      filterName: `All`,
-      filterAmount: 15,
-      filterIsChecked: true
-    },
-    {
-      filterName: `Overdue`,
-      filterAmount: 0,
-      filterIsDisabled: true
-    },
-    {
-      filterName: `Today`,
-      filterAmount: 0,
-      filterIsDisabled: true
-    },
-    {
-      filterName: `Favorites`,
-      filterAmount: 7
-    },
-    {
-      filterName: `Repeating`,
-      filterAmount: 2
-    },
-    {
-      filterName: `Tags`,
-      filterAmount: 6
-    },
-    {
-      filterName: `Archive`,
-      filterAmount: 115
-    }
-  ],
-  cards: [
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `black`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `yellow`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `blue`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `green`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `pink`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `pink`
-    },
-    {
-      cardText: `It is example of repeating task. It marks by wave.`,
-      cardImage: `img/sample-img.jpg`,
-      cardDeadline: {
-        date: `23 September`,
-        time: `11:15 PM`
-      },
-      cardRepeat: `No`,
-      cardHashtag: `#repeat`,
-      cardColor: `pink`
-    }
-  ]
-};
-
-// FILTER
-const filterContainer = document.querySelector(`.main__filter`);
-// Filter Template
-const getFilterElement = dataList.filters.map((filterElement) => {
-  return `
-    <input
-      type="radio"
-      id="filter__${filterElement.filterName.toLowerCase()}"
-      class="filter__input visually-hidden"
-      name="filter"
-      ${filterElement.filterIsChecked ? `checked` : ``}
-	    ${filterElement.filterIsDisabled ? `disabled` : ``}
-    />
-    <label for="filter__${filterElement.filterName.toLowerCase()}" class="filter__label">
-      ${filterElement.filterName} <span class="filter__all-count">${filterElement.filterAmount}</span>
-    </label>
-  `;
-});
-// Render Filters
-const renderFilter = () => {
-  return filterContainer.insertAdjacentHTML(`beforeend`, getFilterElement.join(``));
-};
-renderFilter();
+import * as dataCards from '../data/cards.json';
 
 // BOARD WITH CARDS
 const boardContainer = document.querySelector(`.board__tasks`);
 // Card Template
-const getCardElement = dataList.cards.map((cardElement) => {
+const getCardElement = dataCards.cards.map((cardElement) => {
   return `
     <article class="card card--${cardElement.cardColor} card--repeat">
         <form class="card__form" method="get">
@@ -441,15 +303,5 @@ const getCardElement = dataList.cards.map((cardElement) => {
 const renderCards = () => {
   return boardContainer.insertAdjacentHTML(`beforeend`, getCardElement.join(``));
 };
-renderCards();
 
-// SORT BOARD WITHIN FILTER
-// Filter Elements
-const renderedFilters = document.querySelectorAll(`.filter__label`);
-// Clean Board And Render New Cards
-Array.from(renderedFilters).forEach((renderedFiltersElement) => {
-  renderedFiltersElement.addEventListener(`click`, function () {
-    boardContainer.innerHTML = ``;
-    renderCards();
-  });
-});
+export {boardContainer, renderCards};
