+++ webpack.config.js
@@ -4,7 +4,7 @@ const CopyWebpackPlugin = require(`copy-webpack-plugin`);
 
 module.exports = {
   mode: `development`,
-  entry: `./src/main.js`,
+  entry: `./src/js/main.js`,
   output: {
     filename: `bundle.js`,
     path: path.join(__dirname, `./public`)
